Time: 4 ms (44.28%) | Memory: 33.5 MB (37.94%) - LeetSync

+++ 1001-n-repeated-element-in-size-2n-array/n-repeated-element-in-size-2n-array.cpp
@@ -6,7 +6,7 @@ public:
         for(int &i:nums){
             count[i]++;
         }
-        for(long i=0;i<count.size();i++){
+        for(int i=0;i<count.size();i++){
             if(count[i]==n) return i;
         }
         return -1;
